fix bugs, typos and  improve controller logic and fix response statuses; add multer middleware for file uploads

+++ src/utils/cloudinary.js
@@ -1,6 +1,4 @@
 import { v2 as cloudinary } from 'cloudinary'
-import { log } from 'console';
-import { response } from 'express';
 import fs from 'fs' ;
 
 
@@ -26,7 +24,9 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
         
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove hte locally saved temp files 
+        if(fs.unlinkSync(localFilePath)){
+            fs.unlinkSync(localFilePath);
+        } // remove the locally saved temp files 
         // as the upload operation got failed
         return null;
     }
